fix(seeder): renamed the seeder to be an development seeder
